+++ modules/canvaskit/canvaskit_bindings.cpp
@@ -222,15 +222,6 @@ sk_sp<SkSurface> MakeRenderTarget(sk_sp<GrContext> grContext, SimpleImageInfo si
 // isn't assigned to a JS variable and has delete() called on it.
 // These Apply methods, combined with the smarter binding code allow for chainable
 // commands that don't leak if the return value is ignored (i.e. when used intuitively).
-
-void ApplyAddArc(SkPath& orig, const SkRect& oval, SkScalar startAngle, SkScalar sweepAngle) {
-    orig.addArc(oval, startAngle, sweepAngle);
-}
-
-void ApplyAddOval(SkPath& orig, const SkRect& oval, bool ccw, unsigned start) {
-    orig.addOval(oval, ccw ? SkPathDirection::kCCW : SkPathDirection::kCW, start);
-}
-
 void ApplyAddPath(SkPath& orig, const SkPath& newPath,
                    SkScalar scaleX, SkScalar skewX,  SkScalar transX,
                    SkScalar skewY,  SkScalar scaleY, SkScalar transY,
@@ -243,27 +234,9 @@ void ApplyAddPath(SkPath& orig, const SkPath& newPath,
                                           SkPath::kAppend_AddPathMode);
 }
 
-void ApplyAddRect(SkPath& path, SkScalar left, SkScalar top,
-                  SkScalar right, SkScalar bottom, bool ccw) {
-    path.addRect(left, top, right, bottom, ccw ? SkPathDirection::kCCW : SkPathDirection::kCW);
-}
-
-void ApplyAddRoundRect(SkPath& path, SkScalar left, SkScalar top,
-                  SkScalar right, SkScalar bottom, uintptr_t /* SkScalar*  */ rPtr,
-                  bool ccw) {
-    // See comment below for uintptr_t explanation
-    const SkScalar* radii = reinterpret_cast<const SkScalar*>(rPtr);
-    path.addRoundRect(SkRect::MakeLTRB(left, top, right, bottom), radii,
-                      ccw ? SkPathDirection::kCCW : SkPathDirection::kCW);
-}
-
 void ApplyArcToTangent(SkPath& p, SkScalar x1, SkScalar y1, SkScalar x2, SkScalar y2,
                 SkScalar radius) {
     p.arcTo(x1, y1, x2, y2, radius);
-}
-
-void ApplyArcToOval(SkPath& p, SkRect& oval, SkScalar startAngle, SkScalar sweepAngle, bool forceMoveTo) {
-    p.arcTo(oval, startAngle, sweepAngle, forceMoveTo);
 }
 
 void ApplyArcToArcSize(SkPath& orig, SkScalar rx, SkScalar ry, SkScalar xAxisRotate,
@@ -686,16 +659,6 @@ void drawShapedText(SkCanvas& canvas, ShapedText st, SkScalar x,
 }
 #endif //SK_NO_FONTS
 
-int saveLayerRec(SkCanvas& canvas, const SkPaint* paint,
-                 const SkImageFilter* backdrop, SkCanvas::SaveLayerFlags flags) {
-    return canvas.saveLayer(SkCanvas::SaveLayerRec(nullptr, paint, backdrop, flags));
-}
-
-int saveLayerRecBounds(SkCanvas& canvas, const SkPaint* paint, const SkImageFilter* backdrop,
-                       SkCanvas::SaveLayerFlags flags, const SkRect& bounds) {
-    return canvas.saveLayer(SkCanvas::SaveLayerRec(&bounds, paint, backdrop, flags));
-}
-
 // This is simpler than dealing with an SkPoint and SkVector
 struct PosTan {
     SkScalar px, py, tx, ty;
@@ -996,7 +959,10 @@ EMSCRIPTEN_BINDINGS(Skia) {
         .function("_clipRRect", optional_override([](SkCanvas& self, uintptr_t /* float* */ fPtr, SkClipOp op, bool doAntiAlias) {
             self.clipRRect(ptrToSkRRect(fPtr), op, doAntiAlias);
         }))
-        .function("clipRect", select_overload<void (const SkRect&, SkClipOp, bool)>(&SkCanvas::clipRect))
+        .function("_clipRect", optional_override([](SkCanvas& self, uintptr_t /* float* */ fPtr, SkClipOp op, bool doAntiAlias) {
+            const SkRect* rect = reinterpret_cast<const SkRect*>(fPtr);
+            self.clipRect(*rect, op, doAntiAlias);
+        }))
         .function("_concat", optional_override([](SkCanvas& self, uintptr_t /* SkScalar*  */ mPtr) {
             // See comment above for uintptr_t explanation
             //TODO(skbug.com/10108): make the JS side be column major.
@@ -1004,7 +970,12 @@ EMSCRIPTEN_BINDINGS(Skia) {
             SkM44 m = SkM44::RowMajor(sixteenMatrixValues);
             self.concat(m);
         }))
-        .function("drawArc", &SkCanvas::drawArc)
+        .function("_drawArc", optional_override([](SkCanvas& self, uintptr_t /* float* */ fPtr,
+                                                  SkScalar startAngle, SkScalar sweepAngle,
+                                                  bool useCenter, const SkPaint& paint) {
+            const SkRect* oval = reinterpret_cast<const SkRect*>(fPtr);
+            self.drawArc(*oval, startAngle, sweepAngle, useCenter, paint);
+        }))
         // _drawAtlas takes an array of SkColor. There is no SkColor4f override.
         .function("_drawAtlas", optional_override([](SkCanvas& self,
                 const sk_sp<SkImage>& atlas, uintptr_t /* SkRSXform* */ xptr,
@@ -1032,7 +1003,7 @@ EMSCRIPTEN_BINDINGS(Skia) {
         .function("drawColorInt", optional_override([](SkCanvas& self, SkColor color, SkBlendMode mode) {
             self.drawColor(color, mode);
         }))
-        .function("_drawDRRect",optional_override([](SkCanvas& self, uintptr_t /* float* */ outerPtr,
+        .function("_drawDRRect", optional_override([](SkCanvas& self, uintptr_t /* float* */ outerPtr,
                                                      uintptr_t /* float* */ innerPtr, const SkPaint& paint) {
             self.drawDRRect(ptrToSkRRect(outerPtr), ptrToSkRRect(innerPtr), paint);
         }))
@@ -1041,31 +1012,40 @@ EMSCRIPTEN_BINDINGS(Skia) {
             self.drawDrawable(aImg.get(), x, y);
         }), allow_raw_pointers())
         .function("drawImage", select_overload<void (const sk_sp<SkImage>&, SkScalar, SkScalar, const SkPaint*)>(&SkCanvas::drawImage), allow_raw_pointers())
-        .function("drawImageNine", optional_override([](SkCanvas& self, const sk_sp<SkImage>& image,
-                                                        SkIRect center, SkRect dst,
-                                                        const SkPaint* paint)->void {
-            self.drawImageNine(image, center, dst, paint);
+        .function("_drawImageNine", optional_override([](SkCanvas& self, const sk_sp<SkImage>& image,
+                                                         uintptr_t /* int* */ centerPtr, uintptr_t /* float* */ dstPtr,
+                                                         const SkPaint* paint)->void {
+            const SkIRect* center = reinterpret_cast<const SkIRect*>(centerPtr);
+            const SkRect* dst = reinterpret_cast<const SkRect*>(dstPtr);
+
+            self.drawImageNine(image, *center, *dst, paint);
         }), allow_raw_pointers())
-        .function("drawImageRect", optional_override([](SkCanvas& self, const sk_sp<SkImage>& image,
-                                                        SkRect src, SkRect dst,
-                                                        const SkPaint* paint, bool fastSample)->void {
-            self.drawImageRect(image, src, dst, paint,
-                               fastSample ? SkCanvas::kFast_SrcRectConstraint :
+        .function("_drawImageRect", optional_override([](SkCanvas& self, const sk_sp<SkImage>& image,
+                                                         uintptr_t /* float* */ srcPtr, uintptr_t /* float* */ dstPtr,
+                                                         const SkPaint* paint, bool fastSample)->void {
+            const SkRect* src = reinterpret_cast<const SkRect*>(srcPtr);
+            const SkRect* dst = reinterpret_cast<const SkRect*>(dstPtr);
+            self.drawImageRect(image, *src, *dst, paint,
+                               fastSample ? SkCanvas::kFast_SrcRectConstraint:
                                             SkCanvas::kStrict_SrcRectConstraint);
         }), allow_raw_pointers())
         .function("drawLine", select_overload<void (SkScalar, SkScalar, SkScalar, SkScalar, const SkPaint&)>(&SkCanvas::drawLine))
-        .function("drawOval", &SkCanvas::drawOval)
+        .function("_drawOval", optional_override([](SkCanvas& self, uintptr_t /* float* */ fPtr,
+                                                    const SkPaint paint)->void {
+            const SkRect* oval = reinterpret_cast<const SkRect*>(fPtr);
+            self.drawOval(*oval, paint);
+        }))
         .function("drawPaint", &SkCanvas::drawPaint)
 #ifdef SK_INCLUDE_PARAGRAPH
         .function("drawParagraph", optional_override([](SkCanvas& self, skia::textlayout::Paragraph* p,
-                                                     SkScalar x, SkScalar y) {
+                                                        SkScalar x, SkScalar y) {
             p->paint(&self, x, y);
         }), allow_raw_pointers())
 #endif
         .function("drawPath", &SkCanvas::drawPath)
         // Of note, picture is *not* what is colloquially thought of as a "picture", what we call
         // a bitmap. An SkPicture is a series of draw commands.
-        .function("drawPicture",  select_overload<void (const sk_sp<SkPicture>&)>(&SkCanvas::drawPicture))
+        .function("drawPicture", select_overload<void (const sk_sp<SkPicture>&)>(&SkCanvas::drawPicture))
         .function("_drawPoints", optional_override([](SkCanvas& self, SkCanvas::PointMode mode,
                                                      uintptr_t /* SkPoint* */ pptr,
                                                      int count, SkPaint paint)->void {
@@ -1076,8 +1056,11 @@ EMSCRIPTEN_BINDINGS(Skia) {
         .function("_drawRRect",optional_override([](SkCanvas& self, uintptr_t /* float* */ fPtr, const SkPaint& paint) {
             self.drawRRect(ptrToSkRRect(fPtr), paint);
         }))
-        .function("drawRect", &SkCanvas::drawRect)
-        .function("drawRoundRect", &SkCanvas::drawRoundRect)
+        .function("_drawRect", optional_override([](SkCanvas& self, uintptr_t /* float* */ fPtr,
+                                                    const SkPaint paint)->void {
+            const SkRect* rect = reinterpret_cast<const SkRect*>(fPtr);
+            self.drawRect(*rect, paint);
+        }))
         .function("_drawShadow", optional_override([](SkCanvas& self, const SkPath& path,
                                                      const SkPoint3& zPlaneParams,
                                                      const SkPoint3& lightPos, SkScalar lightRadius,
@@ -1146,18 +1129,11 @@ EMSCRIPTEN_BINDINGS(Skia) {
         .function("restoreToCount", &SkCanvas::restoreToCount)
         .function("rotate", select_overload<void (SkScalar, SkScalar, SkScalar)>(&SkCanvas::rotate))
         .function("save", &SkCanvas::save)
-         // 1 param (only the paint)
-        .function("saveLayer", optional_override([](SkCanvas& self, const SkPaint* p) {
-            return self.saveLayer(nullptr, p);
+        .function("_saveLayer", optional_override([](SkCanvas& self, const SkPaint* p, uintptr_t /* float* */ fPtr,
+                                                     const SkImageFilter* backdrop, SkCanvas::SaveLayerFlags flags)->int {
+            SkRect* bounds = reinterpret_cast<SkRect*>(fPtr);
+            return self.saveLayer(SkCanvas::SaveLayerRec(bounds, p, backdrop, flags));
         }), allow_raw_pointers())
-         // 2 params
-        .function("saveLayer", select_overload<int (const SkRect&, const SkPaint*)>(&SkCanvas::saveLayer),
-                               allow_raw_pointers())
-         // 3 params (effectively with SaveLayerRec, but no bounds)
-        .function("saveLayer", saveLayerRec, allow_raw_pointers())
-         // 4 params (effectively with SaveLayerRec)
-        .function("saveLayer", saveLayerRecBounds, allow_raw_pointers())
-
         .function("scale", &SkCanvas::scale)
         .function("skew", &SkCanvas::skew)
         .function("translate", &SkCanvas::translate)
@@ -1277,7 +1253,12 @@ EMSCRIPTEN_BINDINGS(Skia) {
 
     class_<ShapedText>("ShapedText")
         .constructor<ShapedTextOpts>()
-        .function("getBounds", &ShapedText::getBounds);
+        .function("_getBounds", optional_override([](ShapedText& self,
+                                                     uintptr_t /* float* */ fPtr)->void {
+            SkRect* output = reinterpret_cast<SkRect*>(fPtr);
+            SkRect bounds = self.getBounds();
+            output[0] = bounds;
+        }));
 
     class_<SkFontMgr>("SkFontMgr")
         .smart_ptr<sk_sp<SkFontMgr>>("sk_sp<SkFontMgr>")
@@ -1445,23 +1426,48 @@ EMSCRIPTEN_BINDINGS(Skia) {
         .constructor<const SkPath&>()
         .class_function("_MakeFromCmds", &MakePathFromCmds)
         .class_function("_MakeFromVerbsPointsWeights", &MakePathFromVerbsPointsWeights)
-        .function("_addArc", &ApplyAddArc)
+        .function("_addArc", optional_override([](SkPath& self,
+                                                   uintptr_t /* float* */ fPtr,
+                                                   SkScalar startAngle, SkScalar sweepAngle)->void {
+            // See comment above for uintptr_t explanation
+            const SkRect* oval = reinterpret_cast<const SkRect*>(fPtr);
+            self.addArc(*oval, startAngle, sweepAngle);
+        }))
+        .function("_addOval", optional_override([](SkPath& self,
+                                                   uintptr_t /* float* */ fPtr,
+                                                   bool ccw, unsigned start)->void {
+            // See comment above for uintptr_t explanation
+            const SkRect* oval = reinterpret_cast<const SkRect*>(fPtr);
+            self.addOval(*oval, ccw ? SkPathDirection::kCCW : SkPathDirection::kCW, start);
+        }))
         // interface.js has 3 overloads of addPath
-        .function("_addOval", &ApplyAddOval)
         .function("_addPath", &ApplyAddPath)
         .function("_addPoly", optional_override([](SkPath& self,
-                                                   uintptr_t /* SkPoint* */ pptr,
+                                                   uintptr_t /* SkPoint* */ fPtr,
                                                    int count, bool close)->void {
             // See comment above for uintptr_t explanation
-            const SkPoint* pts = reinterpret_cast<const SkPoint*>(pptr);
+            const SkPoint* pts = reinterpret_cast<const SkPoint*>(fPtr);
             self.addPoly(pts, count, close);
         }))
-        // interface.js has 4 overloads of addRect
-        .function("_addRect", &ApplyAddRect)
-        // interface.js has 4 overloads of addRoundRect
-        .function("_addRoundRect", &ApplyAddRoundRect)
+        .function("_addRect", optional_override([](SkPath& self,
+                                                   uintptr_t /* float* */ fPtr,
+                                                   bool ccw)->void {
+            // See comment above for uintptr_t explanation
+            const SkRect* rect = reinterpret_cast<const SkRect*>(fPtr);
+            self.addRect(*rect, ccw ? SkPathDirection::kCCW : SkPathDirection::kCW);
+        }))
+        .function("_addRRect", optional_override([](SkPath& self,
+                                                   uintptr_t /* float* */ fPtr,
+                                                   bool ccw)->void {
+            self.addRRect(ptrToSkRRect(fPtr), ccw ? SkPathDirection::kCCW : SkPathDirection::kCW);
+        }))
         .function("_addVerbsPointsWeights", &PathAddVerbsPointsWeights)
-        .function("_arcToOval", &ApplyArcToOval)
+        .function("_arcToOval", optional_override([](SkPath& self,
+                                                   uintptr_t /* float* */ fPtr, SkScalar startAngle,
+                                                   SkScalar sweepAngle, bool forceMoveTo)->void {
+            const SkRect* oval = reinterpret_cast<const SkRect*>(fPtr);
+            self.arcTo(*oval, startAngle, sweepAngle, forceMoveTo);
+        }))
         .function("_arcToRotated", &ApplyArcToArcSize)
         .function("_arcToTangent", ApplyArcToTangent)
         .function("_close", &ApplyClose)
@@ -1502,7 +1508,12 @@ EMSCRIPTEN_BINDINGS(Skia) {
 
         .function("setFillType", select_overload<void(SkPathFillType)>(&SkPath::setFillType))
         .function("getFillType", &SkPath::getFillType)
-        .function("getBounds", &SkPath::getBounds)
+        .function("_getBounds", optional_override([](SkPath& self,
+                                                     uintptr_t /* float* */ fPtr)->void {
+            SkRect* output = reinterpret_cast<SkRect*>(fPtr);
+            SkRect bounds = self.getBounds();
+            output[0] = bounds;
+        }))
         .function("computeTightBounds", &SkPath::computeTightBounds)
         .function("equals", &Equals)
         .function("copy", &CopyPath)
@@ -1538,9 +1549,10 @@ EMSCRIPTEN_BINDINGS(Skia) {
 
     class_<SkPictureRecorder>("SkPictureRecorder")
         .constructor<>()
-        .function("beginRecording", optional_override([](SkPictureRecorder& self,
-                                                         const SkRect& bounds) -> SkCanvas* {
-            return self.beginRecording(bounds, nullptr, 0);
+        .function("_beginRecording", optional_override([](SkPictureRecorder& self,
+                                                          uintptr_t /* float* */ fPtr) -> SkCanvas* {
+            SkRect* bounds = reinterpret_cast<SkRect*>(fPtr);
+            return self.beginRecording(*bounds, nullptr, 0);
         }), allow_raw_pointers())
         .function("finishRecordingAsPicture", optional_override([](SkPictureRecorder& self)
                                                                    -> sk_sp<SkPicture> {
@@ -1668,7 +1680,12 @@ EMSCRIPTEN_BINDINGS(Skia) {
 
     class_<SkVertices>("SkVertices")
         .smart_ptr<sk_sp<SkVertices>>("sk_sp<SkVertices>")
-        .function("bounds", &SkVertices::bounds)
+        .function("_bounds", optional_override([](SkVertices& self,
+                                                  uintptr_t /* float* */ fPtr)->void {
+            SkRect* output = reinterpret_cast<SkRect*>(fPtr);
+            SkRect bounds = self.bounds();
+            output[0] = bounds;
+        }))
         .function("uniqueID", &SkVertices::uniqueID);
 
     // Not intended to be called directly by clients
@@ -1839,11 +1856,6 @@ EMSCRIPTEN_BINDINGS(Skia) {
         .field("text",        &ShapedTextOpts::text)
         .field("width",       &ShapedTextOpts::width);
 #endif
-    value_object<SkRect>("SkRect")
-        .field("fLeft",   &SkRect::fLeft)
-        .field("fTop",    &SkRect::fTop)
-        .field("fRight",  &SkRect::fRight)
-        .field("fBottom", &SkRect::fBottom);
 
     value_object<SkIRect>("SkIRect")
         .field("fLeft",   &SkIRect::fLeft)